+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -10,9 +10,9 @@ import java.util.Set;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.model.calendarEvent.EventEnd;
-import seedu.address.model.calendarEvent.EventName;
-import seedu.address.model.calendarEvent.EventStart;
+import seedu.address.model.calendarevent.EventEnd;
+import seedu.address.model.calendarevent.EventName;
+import seedu.address.model.calendarevent.EventStart;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
