+++ src/main/java/seedu/address/model/calendarevent/EventName.java
@@ -1,4 +1,4 @@
-package seedu.address.model.calendarEvent;
+package seedu.address.model.calendarevent;
 
 import static java.util.Objects.requireNonNull;
