+++ src/main/java/seedu/address/model/calendarevent/ReadOnlyCalendarEvent.java
@@ -1,4 +1,4 @@
-package seedu.address.model.calendarEvent;
+package seedu.address.model.calendarevent;
 
 import javafx.beans.property.ObjectProperty;
